+++ einterfaces/mocks/SamlInterface.go
@@ -116,14 +116,14 @@ func (_m *SamlInterface) GetMetadata() (string, *model.AppError) {
 }
 
 // ResetAuthDataToEmail provides a mock function with given fields: includeDeleted, dryRun, specifiedUserIDs
-func (_m *SamlInterface) ResetAuthDataToEmail(includeDeleted bool, dryRun bool, specifiedUserIDs []string) (int64, error) {
+func (_m *SamlInterface) ResetAuthDataToEmail(includeDeleted bool, dryRun bool, specifiedUserIDs []string) (int, error) {
 	ret := _m.Called(includeDeleted, dryRun, specifiedUserIDs)
 
-	var r0 int64
-	if rf, ok := ret.Get(0).(func(bool, bool, []string) int64); ok {
+	var r0 int
+	if rf, ok := ret.Get(0).(func(bool, bool, []string) int); ok {
 		r0 = rf(includeDeleted, dryRun, specifiedUserIDs)
 	} else {
-		r0 = ret.Get(0).(int64)
+		r0 = ret.Get(0).(int)
 	}
 
 	var r1 error
